refactor: add create-keyboard module
