+++ js/index.js
@@ -1,75 +1,7 @@
 import { KEYS_ENG, KEYS_RUS, KEYS_CLASS } from './data.js';
+import { addKeyText, createKeyboard } from './create-keyboard.js';
 
-const body = document.querySelector('body');
-
-const addKeyText = (element, key) => {
-  const theElement = element;
-  theElement.textContent = key;
-};
-
-const createKey = (element, key) => {
-  const elementKey = document.createElement('div');
-  const elementKeyText = document.createElement('span');
-
-  elementKey.classList.add('keyboard__key');
-
-  if (KEYS_CLASS.includes(key)) {
-    if (key === 'Caps Lock') {
-      elementKey.classList.add('keyboard__key_caps-lock');
-    } else {
-      elementKey.classList.add(`keyboard__key_${key.toLowerCase()}`);
-    }
-  }
-
-  if (key === '') {
-    elementKey.classList.add('keyboard__key_space');
-  }
-
-  if (key === '↑') {
-    elementKey.classList.add('keyboard__key_up');
-  }
-
-  if (key === '→') {
-    elementKey.classList.add('keyboard__key_right');
-  }
-
-  if (key === '↓') {
-    elementKey.classList.add('keyboard__key_down');
-  }
-
-  if (key === '←') {
-    elementKey.classList.add('keyboard__key_left');
-  }
-
-  addKeyText(elementKeyText, key);
-
-  element.insertAdjacentElement('beforeend', elementKey);
-  elementKey.insertAdjacentElement('beforeend', elementKeyText);
-};
-
-const createKeyboard = (lang) => {
-  const elementContainer = document.createElement('div');
-  const elementOutput = document.createElement('textarea');
-  const elementKeyboard = document.createElement('div');
-
-  elementContainer.classList.add('container');
-  elementOutput.classList.add('output-text');
-  elementKeyboard.classList.add('keyboard');
-
-  elementOutput.placeholder = '_';
-
-  body.insertAdjacentElement('afterbegin', elementContainer);
-  elementContainer.insertAdjacentElement('beforeend', elementOutput);
-  elementContainer.insertAdjacentElement('beforeend', elementKeyboard);
-
-  lang.forEach((arr) => {
-    const elementKeyboardRow = document.createElement('div');
-    elementKeyboardRow.classList.add('keyboard__row');
-    elementKeyboard.insertAdjacentElement('beforeend', elementKeyboardRow);
-
-    arr.forEach((key) => createKey(elementKeyboardRow, key));
-  });
-};
+// const body = document.querySelector('body');
 
 createKeyboard(KEYS_ENG.lowerCase);
